chore: adjust the structure of geometry folder.
