+++ src/shape/index.ts
@@ -1,0 +1,3 @@
+export * from './base';
+export { default as IntervalShapeFactoty } from './interval';
+export { default as PolygonShapeFactoty } from './polygon';
