+++ src/shape/util/path.ts
