--- tests/unit/geometry/polygon/index-spec.ts
@@ -1,5 +1,0 @@
-describe('polygon', () => {
-  it('polygon', () => {
-    expect('todo').toBe('todo');
-  });
-});
