--- tests/unit/geometry/polygon/shape-spec.ts
@@ -1,5 +1,0 @@
-describe('polygon', () => {
-  it('shape', () => {
-    expect('todo').toBe('todo');
-  });
-});
