+++ tests/unit/shape/interval-spec.ts
